rename `_yaml_data_models` -> `yaml_data_models`
