+++ src/sirocco/core/_tasks/shell_task.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 
 from sirocco.core.graph_items import Task
-from sirocco.parsing._yaml_data_models import ConfigShellTaskSpecs
+from sirocco.parsing.yaml_data_models import ConfigShellTaskSpecs
 
 
 @dataclass(kw_only=True)
