+++ src/sirocco/core/workflow.py
@@ -4,7 +4,7 @@ from itertools import chain, product
 from typing import TYPE_CHECKING, Self
 
 from sirocco.core.graph_items import Cycle, Data, Store, Task
-from sirocco.parsing._yaml_data_models import (
+from sirocco.parsing.yaml_data_models import (
     ConfigBaseData,
     ConfigWorkflow,
 )
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
     from datetime import datetime
     from pathlib import Path
 
-    from sirocco.parsing._yaml_data_models import (
+    from sirocco.parsing.yaml_data_models import (
         ConfigCycle,
         ConfigData,
         ConfigTask,
